update ACI config
